+++ topobank/manager/views.py
@@ -356,7 +356,8 @@ class TopographyCreateWizard(ORCIDUserRequiredMixin, SessionWizardView):
         context['extra_tabs'] = [
             {
                 'title': f"{surface}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{surface.label}'"
@@ -476,7 +477,8 @@ class CorruptedTopographyView(TemplateView):
         context['extra_tabs'] = [
             {
                 'title': f"{surface}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{surface.label}'"
@@ -575,14 +577,16 @@ class TopographyUpdateView(TopographyUpdatePermissionMixin, UpdateView):
         context['extra_tabs'] = [
             {
                 'title': f"{topo.surface.label}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=topo.surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{topo.surface.label}'"
             },
             {
                 'title': f"{topo.name}",
-                'icon': "file-o",
+                'icon': "file",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:topography-detail', kwargs=dict(pk=topo.pk)),
                 'active': False,
                 'tooltip': f"Properties of topography '{topo.name}'"
@@ -671,7 +675,8 @@ class TopographyDetailView(TopographyViewPermissionMixin, DetailView):
         context['extra_tabs'] = [
             {
                 'title': f"{topo.surface.label}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=topo.surface.pk)),
                 'active': False,
                 'login_required': False,
@@ -679,7 +684,8 @@ class TopographyDetailView(TopographyViewPermissionMixin, DetailView):
             },
             {
                 'title': f"{topo.name}",
-                'icon': "file-o",
+                'icon': "file",
+                'icon_style_prefix': 'far',
                 'href': self.request.path,
                 'active': True,
                 'login_required': False,
@@ -721,14 +727,16 @@ class TopographyDeleteView(TopographyUpdatePermissionMixin, DeleteView):
         context['extra_tabs'] = [
             {
                 'title': f"{topo.surface.label}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=topo.surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{topo.surface.label}'"
             },
             {
                 'title': f"{topo.name}",
-                'icon': "file-o",
+                'icon': "file",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:topography-detail', kwargs=dict(pk=topo.pk)),
                 'active': False,
                 'tooltip': f"Properties of topography '{topo.name}'"
@@ -940,7 +948,8 @@ class SurfaceDetailView(DetailView):
         context['extra_tabs'] = [
             {
                 'title': surface.label,
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': self.request.path,
                 'active': True,
                 'login_required': False,
@@ -1037,7 +1046,8 @@ class SurfaceUpdateView(UpdateView):
         context['extra_tabs'] = [
             {
                 'title': f"{surface.label}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{surface.label}'"
@@ -1091,7 +1101,8 @@ class SurfaceDeleteView(DeleteView):
         context['extra_tabs'] = [
             {
                 'title': f"{surface.label}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{surface.label}'"
@@ -1168,7 +1179,8 @@ class SurfaceShareView(FormMixin, DetailView):
         context['extra_tabs'] = [
             {
                 'title': f"{surface.label}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{surface.label}'"
@@ -1297,7 +1309,8 @@ class SurfacePublishView(FormView):
         context['extra_tabs'] = [
             {
                 'title': f"{surface.label}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{surface.label}'"
@@ -1328,7 +1341,8 @@ class PublicationRateTooHighView(TemplateView):
         context['extra_tabs'] = [
             {
                 'title': f"{surface.label}",
-                'icon': "diamond",
+                'icon': "gem",
+                'icon_style_prefix': 'far',
                 'href': reverse('manager:surface-detail', kwargs=dict(pk=surface.pk)),
                 'active': False,
                 'tooltip': f"Properties of surface '{surface.label}'"
@@ -1638,8 +1652,8 @@ class TagTreeView(ListAPIView):
         return tags_for_user(self.request.user, surfaces, topographies).filter(parent=None)
         # Only top level are collected, the children are added in the serializer.
         #
-        # The filtered surfaces and topographies are calculated twice here,
-        # I'm not sure how to circumvent this.
+        # TODO The filtered surfaces and topographies are calculated twice here, not sure how to circumvent this.
+        # Maybe by caching with request argument?
 
     def get_serializer_context(self):
         context = super().get_serializer_context()